moved all authentication function to a class of its own and some other modifications

+++ article.php
@@ -1,19 +1,16 @@
 <?php
-// Create a DB connection
-require './classes/Database.php';
-// access all article functions
-require './classes/Article.php';
+// Classes Autoloader and session start
+require 'includes/init.php';
 
-$db = new Database();
-$conn = $db->getConnection();
+// create a database connection
+$conn = require './includes/db.php';
 
 if (isset($_GET['id'])) {
-    $article = Article::getArticleByID($conn,$_GET['id']);
+    $article = Article::getArticleByID($conn, $_GET['id']);
 } else {
     $article = null;
 }
 ?>
-
 <!-- === START OF HTML DOC ===-->
 <!-- Include header -->
 <?php require './includes/header.php'; ?>
@@ -26,8 +23,12 @@ if (isset($_GET['id'])) {
                 <p class="lead"><?php echo htmlspecialchars($article->content); ?></p>
             </article>
         </li>
-        <a href="./edit_article.php?id=<?php echo $article->id; ?>" class="btn btn-sm btn-primary">Edit</a>
-        <a href="./delete_article.php?id=<?php echo $article->id; ?>" class="btn btn-sm btn-danger">Delete</a>
+        <?php if (Auth::isLoggedIn()) : ?>
+            <a href="./edit_article.php?id=<?php echo $article->id; ?>" class="btn btn-sm btn-primary">Edit</a>
+            <a href="./delete_article.php?id=<?php echo $article->id; ?>" class="btn btn-sm btn-danger">Delete</a>
+        <?php else : ?>
+            <p class="callout-danger">You must be logged in to Edit or Delete. <a href="login.php">Login</a></p>
+        <?php endif ?>
     <?php else : ?>
         <p>Article not found!</p>
         <a href="./index.php" class="btn btn-primary">Back to Home</a>
